feat: simplify mermaid vue component
This component did things mermaid can do on it's own
like providing a temporary container for rendering.

Additionally, it wasn't working with slots as is promoted
in the readme and required for it to be used in the
<mermaid>my diagram</mermaid> syntax.

Oh and ids don't have to be passed in, it's easy to create
a unique one if none is given.